new file for helper functions and styling changes
